refactor(Race, RaceController): 입력 관련 메소드를 RaceController로 분리
Race 클래스에 존재했던 입/출력, 게임 흐름 관련 메소드를 RaceController 클래스로 옮기기 위한 과정입니다
